Perfected the look of the HTML output from the blame module.
This output is a good start and should work fine as a first template that
can be used when implementing HTML output in all other modules.

Once this is done we can work on adding some interactivity. However; it
might be a good idea to just release a version with some simple HTML
output before adding more complex stuff.