Settings for gitinspector can now be set using git-config.
All long options that can be set in gitinspector (with the exception of
--version and --help) can be configured using git-config. This is only the
first step towards the support of git-config. The reading of the command
line also needs to be changed to support overriding settings from
git-config before this is fully usable. Also; support for --exclude is
not yet included.

If we wanted to enable the timeline across all repositories when running
gitinspector we could do;

git-config --global inspector.timeline true

For local settings (per repository); the --local option can of course be
used when calling git-config.